move helper functions into module_helpers folder
