+++ setup.py
@@ -43,9 +43,10 @@ setup(
    test_suite = 'tests',
    package_dir = {
        'vm_network_migration': 'vm_network_migration',
+       'vm_network_migration.module_helpers':'vm_network_migration/module_helpers',
        'vm_network_migration.migrations': 'vm_network_migration/migrations',
        'vm_network_migration.modules':'vm_network_migration/modules'
    },
-   packages=['vm_network_migration', 'vm_network_migration.migrations', 'vm_network_migration.modules'],  #same as name
+   packages=['vm_network_migration', 'vm_network_migration.module_helpers', 'vm_network_migration.migrations', 'vm_network_migration.modules'],  #same as name
    install_requires=install_requires, #external packages as dependencies
 )
