+++ vm_network_migration/modules/subnet_network.py
@@ -93,34 +93,3 @@ class SubnetNetwork():
         auto_mode_status = network_info['autoCreateSubnetworks']
         return auto_mode_status
 
-class SubnetNetworkHelper:
-    def __init__(self, compute, project, zone, region):
-        """ Initialization
-
-        Args:
-            compute: compute engine
-            project: project ID
-            zone: zone of the network
-            region: region of the network
-        """
-        self.compute = compute
-        self.project = project
-        self.zone = zone
-        self.region = region
-
-    def generate_network(self, network, subnetwork):
-        """ Generate a network object
-
-        Args:
-            network: network name
-            subnetwork: subnetwork name
-
-        Returns: a SubnetNetwork object
-
-        """
-        network = SubnetNetwork(self.compute, self.project, self.zone,
-                                self.region, network, subnetwork)
-        network.check_subnetwork_validation()
-        network.generate_new_network_info()
-
-        return network
